Move schema derivation code to module & fix tests
In tests, we often bypass the recipe layer, resulting in views without a
typed SQL schema. Since we still need to be able to form `View` handles
for them, the schema on `View` and `Table` must be optional. The data
structures were already set up for this, but the logic for deriving view
schemas wasn't; it now is.